[Dashboard] Fix tags in form
