+++ frontend/src/features/Dashboard/components/DashboardForm/Toolbar/index.tsx
@@ -5,6 +5,7 @@ import styled from 'styled-components'
 
 import { EditArea } from './EditArea'
 import { DashboardFilters } from './Filters'
+import { FiltersTags } from './Filters/FiltersTags'
 
 import type { DashboardType } from '@features/Dashboard/slice'
 import type { GeoJSON } from 'domain/types/GeoJSON'
@@ -20,14 +21,18 @@ export const Toolbar = forwardRef<HTMLDivElement, ToolbarProps>(
 
     return (
       <Wrapper ref={ref}>
-        <EditArea
-          dashboardKey={key}
-          geometry={geometry}
-          onValidate={geometryToSave => {
-            dispatch(editDashboardArea(geometryToSave, key))
-          }}
-        />
-        <DashboardFilters dashboard={dashboard} />
+        <FirstLine>
+          <EditArea
+            dashboardKey={key}
+            geometry={geometry}
+            onValidate={geometryToSave => {
+              dispatch(editDashboardArea(geometryToSave, key))
+            }}
+          />
+          <DashboardFilters dashboard={dashboard} />
+        </FirstLine>
+
+        <FiltersTags dashboard={dashboard} />
       </Wrapper>
     )
   }
@@ -37,8 +42,13 @@ const Wrapper = styled.div`
   background-color: ${p => p.theme.color.white};
   box-shadow: 0pc 3px 6px #00000029;
   display: flex;
+  flex-direction: column;
   flex-wrap: wrap;
   gap: 16px;
   padding: 16px 24px;
-  align-items: center;
+`
+const FirstLine = styled.div`
+  display: flex;
+  gap: 16px;
+  justify-content: space-between;
 `
